renamed deadline_module.py to addon.py
