+++ server_addon/deadline/client/ayon_deadline/__init__.py
@@ -1,4 +1,4 @@
-from .deadline_module import DeadlineAddon
+from .addon import DeadlineAddon
 from .version import __version__
